deployment/restaurant_service/requirements.txt
asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
drf-yasg==1.21.8
inflection==0.5.1
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pillow==10.4.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
mysql-connector-python==8.0.30

